refactor: ♻️ project detail sidebar social card
