+++ components/AddStar.tsx
@@ -17,6 +17,7 @@
 import { Button, Icon } from "@bbtgnn/polaris-interfacer";
 import { StarFilledMinor, StarOutlineMinor } from "@shopify/polaris-icons";
 import classNames from "classnames";
+import { useAuth } from "hooks/useAuth";
 import useSocial from "hooks/useSocial";
 import useWallet from "hooks/useWallet";
 import { IdeaPoints } from "lib/PointsDistribution";
@@ -28,12 +29,15 @@ const AddStar = ({ id, owner, tiny = false }: { id: string; owner: string; tiny?
   const { likeER, isLiked } = useSocial(id);
   const hasAlreadyStarred = isLiked(id);
   const { t } = useTranslation("common");
+  const { user } = useAuth();
   const { addIdeaPoints } = useWallet();
   const handleClick = async () => {
     await likeER();
     //economic system: points assignments
     addIdeaPoints(owner, IdeaPoints.OnStar);
   };
+
+  if (!user) return null;
 
   return (
     <>
